Create Group Generator
You can modify the name_list variable to contain your own list of names, and the code will generate groups based on the specified group size. If the number of names is not divisible by the group size, the remaining names are put into the last group. The function returns a list of generated groups.